requirements.txt
flask==3.0.0
google-genai==1.0.0
pdf2image==1.17.0
pillow==10.4.0
python-dotenv==1.0.0
opencv-python==4.10.0.84
pytesseract==0.3.10
numpy==1.26.4
weasyprint==62.3
